+++ CHANGELOG.md
@@ -7,5 +7,10 @@
 5. Added minor UI changes for features' title and descriptions
 6. Changed speedtest puffer color to match the theme
 7. Added FPLMN Settings to help users clear forbidden network list
+8. Added QuecWatch's network refresh events to Network Insights
+9. Fixed Custom DNS disable functionality
+10. Added restore previous IMEI to IMEI settings
+11. Added auth-token checker
+12. Added Settings for Live Bandwidth to enable or disable it 
 
 Note: Using https is now recommended